commercetools service class added
