+++ src/main/resources/application-akhildev.properties
@@ -13,3 +13,4 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
 
 secret_auth_key=123
+baseUrl = https://api.europe-west1.gcp.commercetools.com/sunrise-spa/graphql
